thumbnailer: Try harder to create a failed thumbnail
If a failed thumbnail is created because the pixbuf fail to save
(for instance if user is over quota) we should still try to rename
the temporary file that might have been created.
If not, the thumbnail will not be marked as failed, and thumbnailing
will be reattempted.

https://bugzilla.gnome.org/show_bug.cgi?id=728775

origin commit:
https://gitlab.gnome.org/GNOME/gnome-desktop/commit/54f68ab