Revert "[mate-bg] small cleanup"
It is not possible to use the `list` pointer after it has been deleted,
so the "cleanup" this commit made lead to using freed memory if any
item actually got clean up.

This "cleanup" also don't seem meaningful to me, as all it does
otherwise is trade an assignation for a redundant test -- either of
which the compiler might happily optimize out.

This reverts commit 47426c90d10e9f738ecf89f35db94ca8deff55e0.