MateImageMenuItem: Remove unnecessary label field
This was used in GtkImageMenuItem to support stock items, but this is a
deprecated feature that didn't make it to MateImageMenuItem.  Now, this
is entirely unused, and just leads to glue code maintaining an unused
local copy of the label, effectively wasting a little bit of memory and
adding complexity.

Just drop all references to it, without any functional difference.